Change from writing 64-byte packets to dwords.
Having individual device threads write 64-byte packets leads to
uncoalesced memory access. Instead, when writing data of type T,
we split up T objects into 4-byte (dword) parts, which results in
faster writes from the device. On the host side, processing the
data is a bit more involved, since the consecutive dwords of a T
object are now interleaved.

When a wave submits a message, a single wave header is written first,
containing information about the number of lanes involved in writing
data (the active lanes) and the size of the data. Then, for each lane,
there is a lane header, specifiying thread coordinates of the lane.
Finally, there's the actual message data, split up into dwords, and
rounded up to an integral number of dwords.