Add timestamp and source location identifier to wave headers.
Timestamp is taken on the GPU as the very first step in v_submit_message,
and is in clock cycles. The source location identifier is an integer that
can be used to map messages back to the source location from where the
message was submitted. It is up to the dh_comms user to define and use
the mapping.