Implement scalar messages.
All messages start with a wave header, which includes the submission
timestamp, the execution mask for the submitting wave, the IDs of the XCC,
SE and CU on which the wave is executing, and other information
that pertains to the whole wave.

After the wave header follow optional lane headers, which contain the
threadIdx.{x,y,z} values for the active lanes. If wave headers are
included, they are submitted for all active lanes too in case of scalar
messages.

Finally, the message contains optional data. In case of a scalar message,
there is only one data item, and it is submitted by the first active lane.
In case of a vector message, there is one data item for each active lane.

Note: while the test program, heatmap_example.cpp, gives expected results,
a slight modification may cause it to hang if built in Release or RelWithDebInfo
mode. In particular, this happens when we submit a scalar message from the
if-branch in the kernel. We have seen a similar issue before, and it may
or may not be due to the compiler generating incorrect ISA. To be
investigated.