Change message into a class to hide implementation.
Message class currently uses std::vector in its implementation. That
may not be performant if we have to deal with large numbers of messages.
This commit hides the implementation behind an interface that not
depends on std::vector, so that we can easily change out the implementation
later without breaking code that depends on the message class.