Make pass-through behavior of message handlers configurable.
Messages are handled by a chain of message handlers. Each handler
is able to handle one or more message types. The chain may or may
not have multiple handlers for the same message type. For instance,
for memory address messages, we may have one handler that builds a
memory map, and one that does analysis of address coalescing.

If the handlers_pass_through argument to the dh_comms constructor is
true (the default), then all handlers in the chain get to look at the
message, and process it if they can. Otherwise, if handlers_pass_through
is false, then message handling stops with the first handler that can
handle the message type. This is potentially faster if we install many
handlers, and each of them handles a different type.