Add a field 'user_data' to the wave header.
The purpose of the new field is to carry auxiliary data for messages.
For instance, for address messages, the user_data field may encode
whether the address access is a read or a write, the address space
(global vs constant vs LDS), the size of the read/write per lane, etc.

By keeping this kind of auxiliary information in a separate field,
we can keep the user_type field stable while we add capabilities
to message handlers for a given user_type.