Reorganize atomics synchronization.
Previously, both synchronization between multiple waves in device code and
synchronization between host and device code were done using an array of
atomic flags in host-pinned memory, which is not optimal for device-device
synchronization.

This commit changes that to synchronizing device-device in device memory, and
only do host-device synchronization is host-pinned memory. This significantly
increases the overall dh_comms bandwidth.